refactor(mvm-sdk): split lib.rs into builder/ + ctor/ + error + emit
Carve the 743-line lib.rs into per-concern files: errors, lazy
WorkloadBuilder/AppBuilder under builder/, one-shot IR variant
constructors (source/image/entrypoint/resources/network/deps) under
ctor/, and emit on its own. lib.rs shrinks to 74 lines and is now just
crate docs, module declarations, and the prelude re-exports — the
public `use mvm_sdk::*;` surface is unchanged.

No behavior changes; builder_smoke tests + doctest pass, workspace
clippy clean.
